requirements.txt
python-dotenv==1.0.0
rich==13.7.0
kaggle==1.7.4.5
pandas==2.2.3
numpy==2.2.6
requests==2.32.3 